refactor: header layout 하나로 통일
