+++ app/ui/bookshelf/header.tsx
@@ -2,22 +2,14 @@
 
 import { useAtom, useAtomValue } from 'jotai';
 import { useResetAtom } from 'jotai/utils';
-import HeaderLayout from '@/layout/three-items-header';
+import HeaderLayout from '@/layout/header';
 import { AddBook, DeleteBooks, RearrangeBooks } from '@/ui/bookshelf/buttons';
 import { currentModeAtom, selectedItemsAtom } from '@/store/atoms';
 
 type Props = { title: string };
 
 export default function Header({ title }: Props) {
-  return (
-    <HeaderLayout>
-      <ActionButtons />
-      <div>
-        <h2 className="text-xl font-medium dark:text-neutral-200">{title}</h2>
-      </div>
-      <ToggleButtons />
-    </HeaderLayout>
-  );
+  return <HeaderLayout leftItem={<ActionButtons />} title={title} rightItem={<ToggleButtons />} />;
 }
 
 function ToggleButtons() {
